[Feature] delete Artist and Stage
The Stage of Delete Artist and Delete Stage has been filled in and now shows the information. The checker whether or not the stage or Artist can be deleted  has not been implemented yet. the actual deleting has a bug in it that doesn't permanently delete the artist or stage.